se creo clase abstracta y factorizo codigo
