+++ src/main/java/com/oop/examples/patterns/dnd/Armour.java
@@ -1,0 +1,19 @@
+package com.oop.examples.patterns.dnd;
+
+public class Armour {
+  private String type;
+  private int defense;
+
+  public Armour(String type, int defense) {
+    this.type = type;
+    this.defense = defense;
+  }
+
+  public String getType() {
+    return type;
+  }
+
+  public int getDefense() {
+    return defense;
+  }
+}
