+++ src/main/java/com/oop/examples/patterns/dnd/Weapon.java
@@ -1,0 +1,19 @@
+package com.oop.examples.patterns.dnd;
+
+public class Weapon {
+  private String type;
+  private int attack;
+
+  public Weapon(String type, int attack) {
+    this.type = type;
+    this.attack = attack;
+  }
+
+  public String getType() {
+    return type;
+  }
+
+  public int getAttack() {
+    return attack;
+  }
+}
